change file stucture
