--- ai_manager_service/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "ai_manager_service"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-messages = { path = "../messages" }
-sqlx = { workspace = true }
-async-trait = "0.1.68"
-anyhow = "1.0.51"
-chatgpt_rs = "1.0.0"
-tokio = { version = "1.27.0", features = ["full"] }
-eyre = { version = "0.6" }
-
