--- frontend_api/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "frontend_api"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "forntend_api_lib"
-path = "src/lib.rs"
-
-[[bin]]
-name = "frontend_api"
-path = "src/main.rs"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[dependencies]
-# htmx
-maud = { version = "0.26.0", features = ["axum"] }
-axum = { version = "0.7.5", features = ["ws"] }
-askama = { version = "0.12.1", features = ["with-axum"] }
-askama_axum = "0.4.0"
-tower-http = { version = "0.5.1", features = ["fs", "trace"] }
-
-# main
-messages = { path = "../messages" }
-eyre = "0.6.8"
-tokio-tungstenite = { workspace = true }
-anyhow = "1.0.70"
-env_logger = "0.10.0"
-tokio = { version = "1.28.0", features = ["full"] }
-futures-util = "0.3.19"
-futures-channel = "0.3.28"
-futures = "0.3.19"
-serde = { version = "1.0.163", features = ["derive"] }
-serde_json = "1.0.96"
