--- monolith/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "monolith"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-anyhow = { workspace = true }
-sqlx = { workspace = true }
-structopt = { workspace = true }
-twitch_api = { workspace = true }
-
-clap = { workspace = true }
-color-eyre = { workspace = true }
-dotenvy = { workspace = true }
-eyre = { workspace = true }
-futures = { workspace = true }
-#hyper-tls = { workspace = true }
-hyper = { workspace = true }
-reqwest = { workspace = true }
-retainer = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-tokio-tungstenite = { workspace = true }
-tracing = { workspace = true }
-tracing-error = { workspace = true }
-tracing-log = { workspace = true }
-tracing-subscriber = { workspace = true }
-url = { workspace = true }
-messages = { workspace = true }
-ai_manager_service = { workspace = true }
-twitch_listener_service = { workspace = true }
-frontend_api = { workspace = true }
-
-# trying to fix broken dependencies
-twitch_oauth2 = { workspace = true }
-twitch_types = { workspace = true }
-# workaround for https://github.com/twitch-rs/twitch_api/issues/256
-[patch.crates-io.twitch_types]
-git = "https://github.com/twitch-rs/twitch_api"
-[patch.crates-io.twitch_oauth2]
-git = "https://github.com/twitch-rs/twitch_api"
