+++ src/ai_manager/.env
