+++ src/frontend/templates/admin.html
