--- twitch_chat/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "twitch_chat"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-twitch-irc = "5.0.1"
-tokio = { version = "1.32.0", features = ["full"] }
-tracing-subscriber = "0.3"
-tracing = "0.1"
-twitch_message = "0.1.2"
-anyhow = "1.0.75"
-rand = "0.8.5"
-
-# Tui things
-crossterm = "0.27.0"
-anathema = { git = "https://github.com/togglebyte/anathema", branch = "main" }
-tmi = "0.7.1"
