--- twitch_listener_service/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "null_twitch"
-version = "0.1.0"
-edition = "2024"
-
-[[bin]]
-name = "null_twitch"
-path = "src/main.rs"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-sqlx = { workspace = true }
-structopt = "0.3"
-twitch_api = { git = "https://github.com/twitch-rs/twitch_api.git", ref="ce645637a37f10adafd8ffde605b734f10328d40", features = [
-    "reqwest",
-    "helix",
-    "eventsub",
-    "hmac",
-    "unsupported",
-    "mock_api"]}
-clap = { version = "4.1.8", features = ["derive", "env"] }
-color-eyre = { version = "0.6", features = ["capture-spantrace"] }
-dotenvy = "0.15.6"
-eyre = { version = "0.6" }
-futures = "0.3.26"
-hyper = "0.14"
-reqwest = { workspace = true }
-retainer = "0.3.0"
-serde = { version = "1", features = ["derive"] }
-serde_json = { version = "1" }
-tokio = { version = "1.26.0", features = ["macros", "rt-multi-thread"] }
-tokio-tungstenite = { workspace = true }
-tracing = "0.1.37"
-tracing-error = "0.2.0"
-tracing-log = "0.1.3"
-tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
-url = "2.3.1"
-
-
-# workaround for https://github.com/twitch-rs/twitch_api/issues/256
-[patch.crates-io.twitch_types]
-git = "https://github.com/twitch-rs/twitch_api"
-ref = "ef47c46ff0c2f7f8ea66c8291f22ea39351297ce"
